+++ packages/okta-auth-js/test/spec/tokenManager.js
@@ -30,11 +30,21 @@ function setupSync(options) {
 }
 
 describe('TokenManager', function() {
+  let originalLocation;
   beforeEach(function() {
     localStorage.clear();
     sessionStorage.clear();
+
+    // Mock window.location so we appear to be on an HTTPS origin
+    originalLocation = global.window.location;
+    delete global.window.location;
+    global.window.location = {
+      protocol: 'https:',
+      hostname: 'somesite.local'
+    };
   });
   afterEach(function() {
+    global.window.location = originalLocation;
     jest.useRealTimers();
   });
 
@@ -198,7 +208,8 @@ describe('TokenManager', function() {
         'okta-token-storage',
         JSON.stringify({'test-idToken': tokens.standardIdTokenParsed}),
         '2200-01-01T00:00:00.000Z', {
-          sameSite: 'strict'
+          sameSite: 'strict',
+          secure: true
         }
       );
     });
@@ -221,6 +232,7 @@ describe('TokenManager', function() {
         'okta-token-storage',
         JSON.stringify({'test-idToken': tokens.standardIdTokenParsed}),
         '2200-01-01T00:00:00.000Z', {
+          secure: true,
           sameSite: 'strict'
         }
       );
@@ -1223,12 +1235,10 @@ describe('TokenManager', function() {
 
   describe('cookie', function() {
 
-    function cookieStorageSetup(options) {
-      options = options || {};
+    function cookieStorageSetup() {
       return setupSync({
         tokenManager: {
-          storage: 'cookie',
-          secure: options.secure
+          storage: 'cookie'
         }
       });
     }
@@ -1236,19 +1246,6 @@ describe('TokenManager', function() {
     describe('add', function() {
       it('adds a token', function() {
         var client = cookieStorageSetup();
-        var setCookieMock = util.mockSetCookie();
-        client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
-        expect(setCookieMock).toHaveBeenCalledWith(
-          'okta-token-storage',
-          JSON.stringify({'test-idToken': tokens.standardIdTokenParsed}),
-          '2200-01-01T00:00:00.000Z', {
-            sameSite: 'strict'
-          }
-        );
-      });
-
-      it('respects the "secure" option', function() {
-        var client = cookieStorageSetup({ secure: true });
         var setCookieMock = util.mockSetCookie();
         client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
         expect(setCookieMock).toHaveBeenCalledWith(
@@ -1265,27 +1262,50 @@ describe('TokenManager', function() {
 
     describe('get', function() {
       it('returns a token', function() {
-        var client = cookieStorageSetup();
-        client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
-        util.warpToUnixTime(tokens.standardIdTokenClaims.iat);
+        const setCookieMock = util.mockSetCookie();
+        const getCookieMock = util.mockGetCookie(JSON.stringify({
+          'test-idToken': tokens.standardIdTokenParsed
+        }));
+        const client = cookieStorageSetup();
+        util.warpToUnixTime(tokens.standardIdTokenClaims.iat); // token should not be expired
         return client.tokenManager.get('test-idToken')
         .then(function(token) {
           expect(token).toEqual(tokens.standardIdTokenParsed);
+          expect(getCookieMock).toHaveBeenCalledWith('okta-token-storage');
+          expect(setCookieMock).not.toHaveBeenCalled();
         });
       });
 
       it('returns undefined for an expired token', function() {
-        var client = cookieStorageSetup();
+        const setCookieMock = util.mockSetCookie();
+        const getCookieMock = util.mockGetCookie(JSON.stringify({
+          'test-idToken': tokens.standardIdTokenParsed
+        }));
+        const client = cookieStorageSetup();
+        util.warpToUnixTime(tokens.standardIdTokenClaims.exp + 1); // token should be expired
         client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
         return client.tokenManager.get('test-idToken')
         .then(function(token) {
           expect(token).toBeUndefined();
+          expect(getCookieMock).toHaveBeenCalledWith('okta-token-storage');
+          expect(setCookieMock).toHaveBeenCalledWith(
+            'okta-token-storage',
+            '{}',
+            '2200-01-01T00:00:00.000Z', {
+              secure: true,
+              sameSite: 'strict'
+            }
+          );
         });
       });
     });
 
     describe('remove', function() {
       it('removes a token', function() {
+        util.mockGetCookie(JSON.stringify({
+          'test-idToken': tokens.standardIdTokenParsed,
+          'anotherKey': tokens.standardIdTokenParsed
+        }));
         var client = cookieStorageSetup();
         client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
         client.tokenManager.add('anotherKey', tokens.standardIdTokenParsed);
@@ -1295,6 +1315,7 @@ describe('TokenManager', function() {
           'okta-token-storage',
           JSON.stringify({anotherKey: tokens.standardIdTokenParsed}),
           '2200-01-01T00:00:00.000Z', {
+            secure: true,
             sameSite: 'strict'
           }
         );
@@ -1303,6 +1324,10 @@ describe('TokenManager', function() {
 
     describe('clear', function() {
       it('clears all tokens', function() {
+        util.mockGetCookie(JSON.stringify({
+          'test-idToken': tokens.standardIdTokenParsed,
+          'anotherKey': tokens.standardIdTokenParsed
+        }));
         var client = cookieStorageSetup();
         client.tokenManager.add('test-idToken', tokens.standardIdTokenParsed);
         client.tokenManager.add('anotherKey', tokens.standardIdTokenParsed);
@@ -1312,6 +1337,7 @@ describe('TokenManager', function() {
           'okta-token-storage',
           '{}',
           '2200-01-01T00:00:00.000Z', {
+            secure: true,
             sameSite: 'strict'
           }
         );